Refactored to allow more sharing between unrolling components
The only difference between serialize and deserialize at the moment
is one variable type. Made all other code common.
