+++ UnrolledDeserializer.h
@@ -29,13 +29,9 @@ public:
 
 private:
 
-  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> createUnrolled(TClass&, TStreamerInfo&) const;
-  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> createRolled(TClass&, TStreamerInfo&) const;
   TBufferFile bufferFile_;
   TClass* cls_;
-  
   std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> sequences_;
-  
 };
 }
 #endif
