+++ common_unrolling.cc
@@ -4,9 +4,21 @@
 #include "TBaseClass.h"
 #include "TClonesArray.h"
 #include "TStreamerElement.h"
+#include "TStreamerInfo.h"
+
+namespace {
+  TStreamerInfo* buildStreamerInfo(TClass* cl);
+  TStreamerInfo* buildStreamerInfo(TClass* cl, void* pointer);
+  void checkIfCanHandle(TClass* iClass);
+  bool elementNeedsOwnSequence(TStreamerElement* element, TClass* cl);
+  bool canUnroll(TClass* iClass, TStreamerInfo* sinfo);
+
+  std::unique_ptr<TStreamerInfoActions::TActionSequence> setActionSequence(TClass *originalClass, TStreamerInfo *localInfo, TVirtualCollectionProxy* collectionProxy, TStreamerInfoActions::TActionSequence::SequenceGetter_t create, bool isSplitNode, int iID, size_t iOffset );
+
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> buildUnrolledActionSequence(TClass& iClass, TStreamerInfo& iInfo, TStreamerInfoActions::TActionSequence::SequenceGetter_t create);
 
 
-namespace cce::tf::unrolling {
+
   TStreamerInfo* buildStreamerInfo(TClass* cl, void* pointer)
   {
     if (!cl) {
@@ -187,4 +199,110 @@ namespace cce::tf::unrolling {
     return actionSequence;
   }
 
+
+// parent=nullptr id=-1 btype=0 (the default) fStreamerType=-1,  fEntryOffsetLen = 1000 (since TTree::fDefaultEntryOffsetLen is not chagned?) fType = 0 (NOTE: if has custom streamer than fType=-1) fOffset=0 (from default TBranch constructor)
+// if CanSelfReference returns true then SetBit(kBranchAny)
+// Since fStreamerType=-1 then the TLeafElement has fLenType=0 since this is a EDataType::kOther
+//    fFillLeaves = (FillLeaves_t)&TBranchElement::FillLeavesMember;
+//     TStreamerInfoActions::TActionSequence::WriteMemberWiseActionsGetter;
+
+// Now with split level. id=-2 for the 'top level' after top level BranchElement is made, will call Unroll on it
+// ftype is still 0 for the sub items (since we are effectively splitlevel == 1)
+
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> buildUnrolledActionSequence(TClass& iClass, TStreamerInfo& iInfo, TStreamerInfoActions::TActionSequence::SequenceGetter_t create) {
+    std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> vec;
+    vec.reserve(1);
+    vec.emplace_back(setActionSequence(nullptr, &iInfo, nullptr, create, true, -2, 0));
+    TIter next(iInfo.GetElements());
+    TStreamerElement* element = 0;
+    for (Int_t id = 0; (element = (TStreamerElement*) next()); ++id) {
+      if(not elementNeedsOwnSequence(element, &iClass)) {
+        continue;
+      }
+      auto ptr = element->GetClassPointer();
+      if(ptr) {
+        TStreamerInfo* sinfo = buildStreamerInfo(ptr);
+        if(canUnroll(ptr,sinfo)) {
+          auto offset = element->GetOffset();
+          
+          vec.emplace_back(setActionSequence(nullptr, sinfo, nullptr, create, true, -2, offset));
+          
+          TIter next(sinfo->GetElements());
+          TStreamerElement* element = 0;
+          for (Int_t id = 0; (element = (TStreamerElement*) next()); ++id) {
+            if(not elementNeedsOwnSequence(element, &iClass)) {
+              continue;
+            }
+            vec.emplace_back(setActionSequence(nullptr, sinfo, nullptr, create, false, id,offset));
+          }
+          continue;
+        }
+      }
+      //std::cout <<" offset "<<element->GetOffset()<<" "<< (ptr? ptr->GetName(): "?")<<std::endl;
+      vec.emplace_back(setActionSequence(nullptr, &iInfo, nullptr, create, false, id,0));
+    }
+    return vec; 
+  }
+
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> buildActionSequence(TClass& iClass, TStreamerInfoActions::TActionSequence::SequenceGetter_t create) {
+  checkIfCanHandle(&iClass);
+
+  TStreamerInfo* sinfo = buildStreamerInfo(&iClass);
+  if (!sinfo) {
+    //failed to build StreamerInfo
+    abort();
+  }
+  
+  if(canUnroll(&iClass, sinfo) ){
+    return buildUnrolledActionSequence(iClass, *sinfo, create);
+  }
+
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> vec;
+  vec.reserve(1);
+  vec.emplace_back(setActionSequence(nullptr, sinfo, nullptr, create, false, -1, 0));
+  return vec; 
+ }
 }
+
+namespace cce::tf::unrolling {
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> buildReadActionSequence(TClass& iClass) {
+    return buildActionSequence(iClass, TStreamerInfoActions::TActionSequence::ReadMemberWiseActionsGetter);
+  }
+
+  std::vector<std::unique_ptr<TStreamerInfoActions::TActionSequence>> buildWriteActionSequence(TClass& iClass) {
+    return buildActionSequence(iClass, TStreamerInfoActions::TActionSequence::WriteMemberWiseActionsGetter);
+  }
+
+}
+
+
+//NOTE:  TBranchElement::SetFillActionSequence()
+// fInfo
+// fType
+//    for fType = 41
+//     GetInfoImp(), GetParentClass(), GetCollectionProxy()
+//    for fType = 3 or 4 and !fNewIDs.empty()
+//     FindOnfileInfo(fClonesClass, fBranches)
+//        NOTE: only wants fBranches in order to ask each branch for its fInfo
+// types
+//   fType == 3 is TClonesArray element
+//   fType == 4 an STL container that is split
+//   fType == 2 ?? a top level object being unrolled?
+//   fType == 41 is a member data of an object in an unrolled STL container
+/*
+// parent = nullptr fID=-1 fStreamerType=-1 fMother=this fType=0
+std::unique_ptr<TStreamerInfoActions::TActionSequence> UnrolledSerializer::createForContainer(TVirtualCollectionProxy& iProxy) const {
+  auto collProxy = iProxy.Generate(); //fCollProxy
+
+  auto stlType = iProxy.GetCollectionType();
+  if(stlType < 0 ) { stlType = -1*stlType;}
+  return std::unique_ptr<TStreamerInfoActions::TActionSequence>();
+}
+*/
+
+// parent=nullptr id=-1 btype=0 (the default) fStreamerType=-1,  fEntryOffsetLen = 1000 (since TTree::fDefaultEntryOffsetLen is not chagned?) fType = 0 (NOTE: if has custom streamer than fType=-1) fOffset=0 (from default TBranch constructor)
+// if CanSelfReference returns true then SetBit(kBranchAny)
+// Since fStreamerType=-1 then the TLeafElement has fLenType=0 since this is a EDataType::kOther
+//    fFillLeaves = (FillLeaves_t)&TBranchElement::FillLeavesMember;
+//     TStreamerInfoActions::TActionSequence::WriteMemberWiseActionsGetter;
+//
